pyproject.toml
[project]
name = "mood-analyzer"
version = "0.1.0"
description = "Trump mood analysis tool using Truth Social posts"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.51.0",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.12",
    "langchain-huggingface>=0.1.2",
    "langchain-community>=0.3.24",
    "psycopg2-binary>=2.9.10",
    "pgvector>=0.3.6",
    "transformers>=4.51.3",
    "torch>=2.6.0",
    "sentence-transformers>=3.4.1",
    "beautifulsoup4>=4.13.3",
    "python-dotenv>=1.1.0",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "uvicorn>=0.41.0",
    "fastapi>=0.134.0",
]

[project.scripts]
start = "app.main:main"
dev = "app.main:main"
api = "app.api_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "watchdog>=4.0.0",
]
